+++ packages/controllers/adopted-stylesheets/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @phase2/outline-adopted-stylesheets-controller
 
+## 0.0.5
+
+### Patch Changes
+
+- Updates for outline-core-button.
+
 ## 0.0.4
 
 ### Patch Changes